chore: some more tweaks and cleanup
